+++ .gitignore
@@ -1,0 +1,2 @@
+*.html
+html_plot
